refactor: address renderer review feedback
