+++ build.js
@@ -5,6 +5,8 @@ import { writeFile } from 'node:fs/promises'
 const run = async () => {
     try {
         await Promise.all([buildBrowser(), buildModules(), buildDocs()])
+        // The root package is ESM, while buildModules emits .js CommonJS files.
+        // Give the require target its own module-format boundary.
         await writeFile(
             'dist/cjs/package.json',
             '{\n    "type": "commonjs"\n}\n'
