+++ src/state.ts
@@ -1,43 +1,19 @@
 import {
     EffectSubscriber,
-    EffectUnSubscriber,
     StateGetter,
     StateSetter,
     StateSubscriber,
     StateUnSubscriber,
 } from './types.ts'
 import { DoubleLinkedList } from './DoubleLinkedList.ts'
+import {
+    EffectResolver,
+    getCurrentResolver,
+    popCurrentResolver,
+    pushCurrentResolver,
+} from './effect-context.ts'
 
-interface Resolver {
-    sub: StateSubscriber
-    unsubs: DoubleLinkedList<EffectUnSubscriber>
-    children: Resolver[]
-    childCursor: number
-    disposed: boolean
-    clearDependencies: () => void
-    dispose: () => void
-}
-
-const currentResolvers = new DoubleLinkedList<Resolver>()
 const scheduledExecutions = new Set<StateSubscriber>()
-
-/** Run user callbacks without attaching their reads to the active render effect. */
-export const untrack = <T>(callback: () => T): T => {
-    const active: Resolver[] = []
-
-    while (currentResolvers.tail) {
-        active.push(currentResolvers.tail)
-        currentResolvers.pop()
-    }
-
-    try {
-        return callback()
-    } finally {
-        for (let index = active.length - 1; index >= 0; index--) {
-            currentResolvers.push(active[index])
-        }
-    }
-}
 
 let flushPending = false
 
@@ -85,7 +61,7 @@ export const state = <T>(
 
     return Object.freeze([
         () => {
-            const currentResolver = currentResolvers.tail
+            const currentResolver = getCurrentResolver()
             if (
                 typeof currentResolver?.sub === 'function' &&
                 !subs.has(currentResolver.sub)
@@ -126,10 +102,10 @@ export const effect = <T>(sub: EffectSubscriber<T>) => {
     let isRunning = false
     let pendingReRun = false
 
-    const parent = currentResolvers.tail
+    const parent = getCurrentResolver()
     const childIndex = parent ? parent.childCursor++ : -1
 
-    const res: Resolver = {
+    const res: EffectResolver = {
         sub: () => run(),
         unsubs: new DoubleLinkedList(),
         children: [],
@@ -170,14 +146,14 @@ export const effect = <T>(sub: EffectSubscriber<T>) => {
         isRunning = true
         res.clearDependencies()
         res.childCursor = 0
-        currentResolvers.push(res)
+        pushCurrentResolver(res)
 
         try {
             value = sub(value)
         } catch (e) {
             console.error(e)
         } finally {
-            currentResolvers.pop()
+            popCurrentResolver()
 
             const staleChildren = res.children.splice(res.childCursor)
             for (const child of staleChildren) child.dispose()
